+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2021-09-10
+
+### Added
+
+Add classes Authentication Controller and Dashboard Controller.
+
+### Changed
+
+- Change handler for the router in the index file.
+
 ## [1.0.1] - 2021-08-15
 
 ### Fixed
@@ -24,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 **first commit**
 
-[unreleased]: https://github.com/ManuelGil/project-skeleton/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/ManuelGil/project-skeleton/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/ManuelGil/project-skeleton/compare/v1.0.1...v2.0.0
 [1.0.1]: https://github.com/ManuelGil/project-skeleton/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ManuelGil/project-skeleton/compare/v0.1.0...v1.0.0
 [0.1.0]: https://github.com/ManuelGil/project-skeleton/releases/tag/v0.1.0